update using models class
