+++ controllers/library_crud.js
@@ -3,78 +3,44 @@ const ObjectID = require('mongodb').ObjectID
 // Connection URL
 const url = 'mongodb://localhost:27017/library';
 
+// require model
+const ModelLibrary = require('../models/library')
+const Library = new ModelLibrary()
+
 //insert data
 let insertData = (req, res) => {
-  MongoClient.connect(url, function(err, db) {
-    if(err){
-      res.status(500).send(err)
-    } else {
-      let obj = {
-          isbn: req.body.isbn,
-          title: req.body.title,
-          author: req.body.author,
-          catagory: req.body.catagory,
-          stock: req.body.stock 
-        }
-      
-      // Insert a single document
-      db.collection('books').insertOne(obj, function(err, result) {
-        db.close();
-        res.send(result)
-      })
-      
-    }
+  Library.insertData(req, (err, data)=>{
+    if(err) res.status(501).send(err)
+    else res.send(data)
   })
+
 }
 
 //get all data
 let getAll = (req, res) => {
-  MongoClient.connect(url, function(err, db) {
-    if(err){
-      res.status(500).send(err)
-    } else {
-      db.collection('books').find({}).toArray((err, docs)=>{
-        db.close();
-        res.send(docs)
-      })
-    }
-    
+  Library.getAll((err, data)=>{
+    if(err) res.status(501).send(err)
+    else res.send(data)
   })
+
 }
 
 //edit data
 let editData = (req, res) => {
-  MongoClient.connect(url, function(err, db) {
-    let obj = {
-        isbn: req.body.isbn,
-        title: req.body.title,
-        author: req.body.author,
-        catagory: req.body.catagory,
-        stock: req.body.stock 
-      }
-    
-    var myquery = { _id: new ObjectID(req.params.id) };
-    var newvalues = obj
-    
-    db.collection("books").updateOne(myquery, newvalues, function(err, result) {
-      if (err) throw err;
-      db.close();
-      res.send({result})
-    });
-    
+  Library.editData(req, (err, data)=>{
+    if(err) res.status(501).send(err)
+    else res.send(data)
   })
+
 }
 
 //delete data
 let deleteData = (req, res) => {
-  MongoClient.connect(url, function(err, db) {
-    var myquery = { _id: new ObjectID(req.params.id) };
-    db.collection("books").deleteOne(myquery, function(err, result) {
-      db.close();
-      res.send(result)
-    })
-    
+  Library.deleteData(req, (err, data)=>{
+    if(err) res.status(501).send(err)
+    else res.send(data)
   })
+
 }
 
 module.exports = {
